update project structure
