+++ .gitignore
@@ -5,3 +5,13 @@ build_k8s
 .terraform*
 secret.auto.tfvars
 *.tfstate*
+
+# Binaries for programs and plugins
+*.exe
+*.exe~
+
+# Test binary, build with `go test -c`
+*.test
+
+# Output of the go coverage tool, specifically when used with LiteIDE
+*.out
